+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.4.0 2017-04-02
 - fix `autoIncrement`
+- update deps
+- update eslint config, use no semi
 
 ## v0.3.1 2016-11-18
 - map all `CURRENT_TIMESTAMP` to `() => new Date`